Refactor into multipage Streamlit app
